Split sound functions to Audible class.
